Moved session connection logic into a hook
