+++ client/src/screens/GameScreen.tsx
@@ -1,53 +1,18 @@
-import { useEffect } from "react";
 import { GameMap } from "../components/GameMap";
 import { GenRandomNodeButton } from "../components/GenRandomNodeButton";
-import { useLocation } from "../hooks/useLocation";
 import { useUserStore } from "../stores/userStore";
 import { ScreenHandler } from "../types";
-import { useNodeStore } from "../stores/nodeStore";
 import { LogoutButton } from "../components/LogoutButton";
 import { useSessionStore } from "../stores/sessionStore";
-import { connectSocket, disconnectSocket, setLocation } from "../utils/sockets";
+import { useSessionConnection } from "../hooks/useSessionConnection";
 
 
 export function GameScreen({ setScreen }: { setScreen: ScreenHandler }) {
-  const { position } = useLocation();
+  useSessionConnection();
 
   const username = useUserStore((state) => state.username);
   const userId = useUserStore((state) => state.userId);
-
-  const fetchNodes = useNodeStore((state) => state.fetchNodes);
-
-  if (!username) {
-    console.error("Username not valid while GameScreen is active");
-    setScreen("login");
-  }
-
-  useEffect(() => {
-    console.warn("Calling connectSocket")
-    connectSocket();
-    return () => {
-      console.warn("Calling disconnectSocket in cleanup")
-      disconnectSocket();
-    }
-  }, []);
-
-  // Update stored position when position changes
-  useEffect(() => {
-    // Don't update if it wasn't null and is now?
-    // console.warn("position update triggered useEffect");
-    if (!position) {
-      console.warn("\t!position. skipping");
-      return;
-    }
-    // console.log("\tcalling setLocation");
-    setLocation(position);
-  }, [position]);
-
-  // Fetch nodes when position changes (this effectively is a timer)
-  useEffect(() => {
-    fetchNodes(position);
-  }, [position, fetchNodes]);
+  const position = useSessionStore((state) => state.location);
 
   if (!position) {
     return <h2 className="text-xl font-bold m-2">Locating Position...</h2>;
